Remove DOM library hard dependency
Backbone only has a dependency on a DOM library if you intend to use
Backbone.View. If you don't include a DOM library, `Backbone.$` will be
undefined.

Since Backbone.localStorage doesn't have a hard dependency on
`jQuery.Deferred`, it should check for the existence of `Backbone.$`
before accessing one of its members.

This makes it possible to use Backbone.localStorage with Backbone and
an alternative to Backbone.View.